add checkVault()
