+++ scripts/deploySafeKeep.js
@@ -1,0 +1,56 @@
+// We require the Hardhat Runtime Environment explicitly here. This is optional
+// but useful for running the script in a standalone fashion through `node <script>`.
+//
+// When running the script with `hardhat run <script>` you'll find the Hardhat
+// Runtime Environment's members available in the global scope.
+const { defaultAccounts } = require("ethereum-waffle");
+const hre = require("hardhat");
+
+async function main() {
+  // Hardhat always runs the compile task when running scripts with its command
+  // line interface.
+  //
+  // If this script is run directly using `node` you may want to call compile
+  // manually to make sure everything is compiled
+  // await hre.run('compile');
+
+  // We get the contract to deploy
+  const Greeter = await hre.ethers.getContractFactory("SafeKeep");
+  const greeter = await Greeter.deploy();
+
+  [
+    "0x90F79bf6EB2c4f870365E785982E1f101E93b906",
+    "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65",
+    "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc",
+    "0x976EA74026E726554dB657fA54763abd0C3a0aa9",
+  ];
+
+  await greeter.deployed();
+
+  console.log("Greeter deployed to:", greeter.address);
+  //console.log(greeter);
+
+  await greeter.createVault(
+    [
+      "0x90F79bf6EB2c4f870365E785982E1f101E93b906",
+      "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65",
+      "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc",
+      "0x976EA74026E726554dB657fA54763abd0C3a0aa9",
+    ],
+    100000,
+    "0x90F79bf6EB2c4f870365E785982E1f101E93b906",
+    { value: 100000 }
+  );
+  const firstvault = await greeter.getAllInheritors(0);
+  //console.log(firstvault);
+  console.log(await greeter.checkVault(0));
+}
+
+// We recommend this pattern to be able to use async/await everywhere
+// and properly handle errors.
+main()
+  .then(() => process.exit(0))
+  .catch((error) => {
+    console.error(error);
+    process.exit(1);
+  });
